refactor: MyInfoPage 이름 수정
