+++ frontend/src/router.tsx
@@ -9,14 +9,13 @@ import SeeAllNearTopics from './pages/SeeAllNearTopics';
 import SeeAllLatestTopics from './pages/SeeAllLatestTopics';
 import SeeTogetherTopics from './pages/SeeTogetherTopics';
 import Favorite from './pages/Favorite';
-import Profile from './pages/Profile';
 import KakaoRedirectPage from './pages/KaKaoRedirectPage';
 import ErrorPage from './pages/ErrorPage';
 import AskLoginPage from './pages/AskLoginPage';
 import { ReactNode } from 'react';
 import AuthLayout from './components/Layout/AuthLayout';
 import NotFound from './components/NotFound';
-import MyInfoPage from './pages/MyInfoPage';
+import Profile from './pages/Profile';
 import LoginError from './components/NotFound/LoginError';
 
 interface routeElement {
@@ -83,7 +82,7 @@ const routes: routeElement[] = [
       },
       {
         path: 'my-page',
-        element: <MyInfoPage />,
+        element: <Profile />,
         withAuth: true,
       },
       {
